Use correct StorageIndex in SparseMatrix
From eigen changelog about SparseMatrix::Index:
In Eigen 3.2, this type was not documented and it was improperly defined as the storage index type (e.g., int), whereas it is now deprecated and always defined as Eigen::Index. Codes making use of SparseMatrix::Index, might thus likely have to be changed to use SparseMatrix::StorageIndex instead.

Former-commit-id: a1365993624d7c9157a0e24505b607c5ba626dd2