Allow use of external Eigen3 build indicated by Eigen3_DIR
In projects where Eigen3 is built externally (for other dependencies also linked in),
it is preferrable to use the same source. Specifying the CMake variable Eigen3_DIR
at configuration will trigger a search for Eigen3 in that directory. If successful,
Eigen will not be downloaded as a dependency. If unsuccessful, the download will still
occur.  If Eigen3_DIR is blank, no search occurs.